Given up

+++ src/vm/page.h
@@ -17,7 +17,7 @@ struct page
     //TODO - remove the name
     //const char *name;           /* Name of the page if stored in filesys. */
     struct file *file;          /* File to lazily load the page from. */
-    int file_start_pos;       /* Starting position in the file to read
+    int16_t file_start_pos;       /* Starting position in the file to read
                                    the page from. */
     uint16_t file_read_bytes;        /* How many bytes to read from the file. */
     bool write;                 /* Indication of read/write permissions. */
